create global fixture
This way the buffers get allocated and cleaned up for each test case and
the buffered output gets shown in syslog and console.

The custom main approach could also work, but the buffer cleanup could
get really complicated.

Also io_context must be on top of the class variables to avoid any
errors.